refactor: hoist shared modality logic out of the profile module
Metric inference, distinct-frame normalization, measure phrasing, and the
formatting helpers for titles, per-measure sections, function headings, and
diff sections were all in the profile module, though none of them is specific
to sampled call stacks. They move to `src/modalities/`, along with the
Markdown table helpers the format tests share.
