+++ src/formats/pprof/parse.ts
@@ -1,11 +1,11 @@
 import { Profile as PprofProto } from 'pprof-format'
+import { determineMetric } from '../../modalities/metric.ts'
+import type { Metric } from '../../modalities/metric.ts'
 import type {
   Profile,
   ProfileStackFrame,
   Sample,
 } from '../../modalities/profile/index.ts'
-import { determineMetric } from '../../modalities/profile/metric.ts'
-import type { Metric } from '../../modalities/profile/metric.ts'
 
 export const parsePprof = (bytes: Uint8Array): Profile[] => {
   const profile = PprofProto.decode(bytes)
